use double instead of int for progress and enable it
git-svn-id: svn+ssh://mono-cvs.ximian.com/svn/monodevelop/trunk@1194 b5faff8f-5bce-0310-b493-a43f95cdeebf

+++ MonoDevelop/src/Main/Base/Gui/IProgressMonitor.cs
@@ -13,9 +13,9 @@ namespace MonoDevelop.Gui
 	/// </summary>
 	public interface IProgressMonitor 
 	{
-		void BeginTask(string name, int totalWork);
+		void BeginTask (string name, int totalWork);
 		
-		void Worked(int work, string status);
+		void Worked (double work, string status);
 		
 		void Done();
